dockerfiles/mef-soa/requirements.txt
fastapi[all]
docker
httpx
authlib
python-jose